+++ packaging/fedora/setup/engine-upgrade.py
@@ -1054,7 +1054,8 @@ def deleteEngineSysconfig():
 def setupVarPrivileges():
     # previous versions mixed root/ovirt
     # ownership in these directories
-    shutil.rmtree(basedefs.DIR_ENGINE_TMP)
+    if os.path.exists(basedefs.DIR_ENGINE_TMP):
+        shutil.rmtree(basedefs.DIR_ENGINE_TMP)
     utils.execCmd(
         cmdList=('chown', '-Rh', 'ovirt:ovirt', basedefs.DIR_DEPLOYMENTS),
         failOnError=True,
